Implement settings. (#45)

+++ script.js
@@ -1,17 +1,34 @@
-const noteForm = $("#note-form");
-const noteInput = $("#note-input");
-const noteList = $("#note-list");
-
+// All global variables which are in local storage.
 let notes = {};
 let dates = {};
 let id = 0;
+let dateTime = true;
+let confirmDelete = true;
 
+// Global variables not in local storage.
+let curEdit = [];
+
+// Retrieve all the varibales from browser storage.
 if (localStorage.getItem("notes") !== null) {
   notes = JSON.parse(localStorage.getItem("notes"));
+  id = Number(localStorage.getItem("id"));
   if (localStorage.getItem("dates") !== null) {
     dates = JSON.parse(localStorage.getItem("dates"));
   }
-  id = Number(localStorage.getItem("id"));
+  if (localStorage.getItem("dateTime") !== null) {
+    if (localStorage.getItem("dateTime") == "true") {
+      dateTime = true;
+    } else {
+      dateTime = false;
+    }
+  }
+  if (localStorage.getItem("confirmDelete") !== null) {
+    if (localStorage.getItem("confirmDelete") == "true") {
+      confirmDelete = true;
+    } else {
+      confirmDelete = false;
+    }
+  }
 }
 
 // Add notes after browser refresh.
@@ -19,9 +36,9 @@ $.each(notes, function (curId) {
   addNoteToUl(curId);
 });
 
-let curEdit = [];
-
+// Function to add the notes to list.
 function addNoteToUl(curId) {
+  // Initializing the list element and the text that goes in it.
   const li = $("<li>").attr("id", "li-" + curId);
   const p = $("<p>")
     .attr("id", "p-" + curId)
@@ -43,14 +60,21 @@ function addNoteToUl(curId) {
     .addClass("delete-button")
     .attr("id", "delete-" + curId);
 
-  // Adding the date to the note.
+  // Initializing the date element for the note.
   let curDate;
-  if (dates[curId] !== undefined) {
-    curDate = new Date(dates[curId]).toLocaleString();
+
+  // Calculations for what date to add to the note.
+  if (dateTime) {
+    if (dates[curId] !== undefined) {
+      curDate = new Date(dates[curId]).toLocaleString();
+    } else {
+      // Set default date in case no date is available due to migration from an old to a new version of the site.
+      dates[curId] = new Date();
+      curDate = dates[curId].toLocaleString();
+      localStorage.setItem("dates", JSON.stringify(dates));
+    }
   } else {
-    dates[curId] = new Date();
-    curDate = dates[curId].toLocaleString();
-    localStorage.setItem("dates", JSON.stringify(dates));
+    curDate = new Date(dates[curId]).toLocaleDateString();
   }
 
   const date = $("<p>")
@@ -62,7 +86,7 @@ function addNoteToUl(curId) {
   const editButtonIcon = $("<i>").addClass("fas fa-edit");
   const deleteButtonIcon = $("<i>").addClass("fas fa-trash-alt");
 
-  // Adding the note to the html with the buttons included.
+  // Adding the li element to note with all it's elements.
   editButton.append(editButtonIcon);
   deleteButton.append(deleteButtonIcon);
   div1.append(date).append(editButton).append(deleteButton);
@@ -70,61 +94,75 @@ function addNoteToUl(curId) {
   li.append(div2);
   li.append(p);
   li.append(textarea);
-  noteList.append(li);
+  $("#note-list").append(li);
 
-  // Clear input
-  noteInput.val("");
+  // Clear input.
+  $("#note-input").val("");
 }
 
 function addNote() {
-  // Get note text
-  const text = noteInput.val().trim();
+  // Get input text.
+  const text = $("#note-input").val().trim();
   if (text === "") {
     return;
   }
 
-  //Increment ID
+  // Update all the variables.
   id += 1;
-
-  // Add note to dictionary
   notes[id] = text;
   dates[id] = new Date();
 
   addNoteToUl(id);
 
-  //Store the notes in browser storage
+  // Update the variables in browser storage.
   localStorage.setItem("dates", JSON.stringify(dates));
   localStorage.setItem("notes", JSON.stringify(notes));
   localStorage.setItem("id", id);
 }
 
-noteForm.on("submit", function (event) {
+// Event listener for addition of new notes.
+$("#note-form").on("submit", function (event) {
   event.preventDefault();
   addNote();
 });
 
-// Delete note.
+// Event listener for deleting a note.
 $(document).on("click", ".delete-button", function () {
+  // Check if confirm delete is true.
+  if (confirmDelete) {
+    const confirmDialog = confirm("Are you sure you want to delete this note?");
+    if (!confirmDialog) {
+      return;
+    }
+  }
+
   const del_id = $(this).attr("id");
   const liId = "#li-" + del_id.substr(7);
+
+  // Deleting all the required elements and variables.
   $(liId).remove();
   delete notes[del_id.substr(7)];
   delete dates[del_id.substr(7)];
+
+  // Update the deleted variables in browser storage.
   localStorage.setItem("notes", JSON.stringify(notes));
   localStorage.setItem("dates", JSON.stringify(dates));
 });
 
+// Event listener for editing a note.
 $(document).on("click", ".edit-button", function () {
   const edit_id = $(this).attr("id");
   const id = edit_id.substr(5);
   const textareaId = "#textarea-" + id;
   const pId = "#p-" + edit_id.substr(5);
 
+  // Keep track of all the notes that in an edit state.
   if (curEdit.indexOf(Number(id)) == -1) {
     $(pId).hide();
     $(textareaId).val(notes[id]).show();
     curEdit.push(Number(id));
   } else {
+    // Update the note with the edited text.
     notes[id] = $(textareaId).val().trim();
     $(pId).show().text(notes[id]);
     $(textareaId).val("").hide();
@@ -132,6 +170,7 @@ $(document).on("click", ".edit-button", function () {
     curEdit.splice(remIndex, 1);
   }
 
+  // Update the variables in browser storage.
   localStorage.setItem("notes", JSON.stringify(notes));
 });
 
@@ -152,13 +191,14 @@ $(".no-tabs").on("keydown", function (e) {
   }
 });
 
+// Event listener for handling switching between the setting and the home page.
 $(document).ready(function () {
   const noteElements = $("#note-elements");
-  const settingsElements = $("#settings-elements");
+  const settingsElements = $("#setting-container");
   const settingsButton = $("#settings-button");
   const homeButton = $("#home-button");
 
-  // Show note elements and hide settings
+  // Show note elements and hide settings.
   function showNoteElements() {
     noteElements.show();
     settingsElements.hide();
@@ -166,7 +206,7 @@ $(document).ready(function () {
     homeButton.hide();
   }
 
-  // Show settings and hide note elements
+  // Show settings and hide note elements.
   function showSettings() {
     noteElements.hide();
     settingsElements.show();
@@ -176,6 +216,7 @@ $(document).ready(function () {
 
   // Handle home button click
   homeButton.on("click", function () {
+    window.location.reload();
     showNoteElements();
   });
 
@@ -187,3 +228,57 @@ $(document).ready(function () {
   // Show note elements by default
   showNoteElements();
 });
+
+// DateTime Settings.
+const dayDateSetting = $("#dayDate-setting");
+let settingState = true;
+
+dayDateSetting.on("change", function () {
+  settingState = $(this).is(":checked");
+  if (settingState) {
+    dateTime = false;
+    localStorage.setItem("dateTime", false);
+  } else {
+    dateTime = true;
+    localStorage.setItem("dateTime", true);
+  }
+});
+
+// Confirm note delete Settings.
+const confirmDeleteSetting = $("#confirmDelete-setting");
+let confirmDeleteState = true;
+
+confirmDeleteSetting.on("change", function () {
+  confirmDeleteState = $(this).is(":checked");
+  if (confirmDeleteState) {
+    confirmDelete = true;
+    localStorage.setItem("confirmDelete", true);
+  } else {
+    confirmDelete = false;
+    localStorage.setItem("confirmDelete", false);
+  }
+});
+
+// Clear data Settings.
+const clearDataButton = $("#clear-data-button");
+
+clearDataButton.on("click", function () {
+  // Show confirmation dialog
+  showConfirmationDialog_dataDeletion();
+});
+
+// Confirmation dialog for deletion of all notes.
+function showConfirmationDialog_dataDeletion() {
+  const confirmDialog = confirm(
+    "Are you sure you want to clear all site data?"
+  );
+  if (confirmDialog) {
+    dayDateSetting[0].checked = false;
+    confirmDeleteSetting[0].checked = true;
+    localStorage.clear();
+    window.location.reload();
+    console.log("Clearing site data...");
+  } else {
+    return;
+  }
+}
